Added solution - LeetHub

+++ Count of Subarrays - GFG/count-of-subarrays.cpp
@@ -14,7 +14,7 @@ public:
 
 	ll countSubarray(int arr[], int n, int k) 
 	{
-	    ll count=0;
+	   int count=0;
 	   int j=0;
 	   for(int i=0;i<n;i++)
 	   {
